state-sync: don't request trie nodes to peers known not to have them
Keep track of which nodes we requested to every peer so that we know
which ones they're missing when they reply with just a subset of what
was requested. That way we can avoid re-requesting those missing nodes
to peers that don't have them. #1073

Also, when a reply from a peer contains only a subset of what was
requested, we immediately re-request the missing ones. Closes: #1077