Enable header peer catchup
When a new peer is selected to sync headers, it may start from far
behind, if there is a large backlog of bodies being downloaded
during fast sync. We were re-requesting the same headers from them
in an infinite loop (only resolved when we disconnect from the peer
and try again, after taming the backlog).