Ensure ethstats plugin uses a timeout to wait on PeerCount
Previously, the Ethstats Plugin waited on the PeerCount
forever. But since the PeerPool is not ready by the time
that the Ethstats plugin sends the request, it just hangs
there waiting and at some point the websocket connection
times out.

This change introduces a timeout of 500ms for the PeerPool
to answer. If the PeerPool does not answer within 500ms,
we assume 0 peers and move on.

Fixes #1379