tests: include missing Constantinople tests.
2 COMMITS SQUASHED:

fixtures: include missing Constantinople in helpers.

Debugging CREATE2 using the full "Blockchain" tests in `fixtures`
(they are disabled in `py-evm`, because they're slow and a duplication
of "State" tests).

A few definitions are missing - so add them.

tests/conftest: fix VM-tracing log-to-file helper.

Also move `import`s out of the helper, so they're not encountered
by the interpreter on every invocation of the helper.