tests: add more tests to INCORRECT_UPSTREAM_TESTS.
One is existing `RevertInCreateInInit`, but now for Constantinople,
not just Byzantium.

The other is `RevertInCreateInInitCreate2`, which contains the same
"synthhetic" state which can't be arrived at by regular means in the
EVM. It's likely a copy-paste atavism.

The rest are variants testing specifically for what happens on
collisions with a state like above. It has already been estabilished
that there are two schools of thought: one with `geth` and `aleth`
(and `testeth` that generates the tests); the other with `py-evm` and
`parity`.