eth/vm/state: recurse collect_touched_accounts() always, unless computation had error.
Previously, "nested" computations would be recursed into even if
they had an error; whereas "origin" computation would be recursed into
on success only.

This commit effectively makes so that "nested" comp-ns will be recursed
into only if they are successful (that is, same logic as for "origin"
comp-n).

The rest of the changes are comments, heredoc brush-ups, and similar
fluff from the debugging session.