EIP-161 collect_touched_accounts(): geth-style
Plus a bonus fix:

tests: fix import when DEBUG2 logging.

Logging utilities have moved to separate package, eth_utils.

===

The EIP-161 implementation evolved over a few commits originally,
here is the history:

Geth is now filling (some) tests in upstream/tests, so for edge-cases
that are impossible on mainnet, prefer geth's approach.

Walk the entire computation tree, including branches that errored
out (got REVERTed, OOGed, etc.), looking for wabbit RIPEMD160,
_always_ collecting it. Skip everything else on errored branches.

---

Don't collect if nested computation errored.

In the general non-RIPEMD160 case, if the computation errored, it
would still be collected, because of the computation.is_origin_computation
condition.

---

collect_touched_accounts(): clean up.

Now that `computation.is_origin_computation` no longer needs to
be considered, it's fine to move the RIPEMD160 special-casing
back into the `if` clause, because the condition is the same.

---

collect_touched_accounts():
remove computation.is_origin_computation when considering SELFDESTRUCT beneficiaries + refactor.

`computation.is_origin_computation` has been present since the file was first introduced by @pipermerriam:

https://github.com/ethereum/py-evm/commit/69a8385476983b4a060f307061ab258027b64eff#diff-28f7a9ef4182d19a95f4cccb1c073ef3

However, whether it's present or not has no effect on the tests passing.

(Possibly, this is due to the case not being covered in the tests...)

I can't find a place which explicitly states that beneficiary collection
should only be skipped if the error happened in the origin computation.

From the fact that having a `computation.is_origin_computation` in the
next if-clause (`if computation.msg.to != constants.CREATE_CONTRACT_ADDRESS`)
caused an error in a RevertPrecompiledTouchExactOOG test, it would seem
that upstream (geth v1.9.6) does not consider this a necessary condition.

Therefore, removing the condition here, too; and re-working the if-else
clauses to look like they did previously (moving the RIPEMD160 special-casing
as a sub-clause of the main clause).