Erase any 'locked' account writes on deletion
The bug was: when a previous transaction creates/modifies an account,
and then is "locked", those pending changes are written even if the
account is deleted is a subsequent transaction.

The fix was: clear out all locked storage changes when you lock in an
account deletion.

This requires carefully managing reverts that cross over to the
pre-delete state. So we track pending writes in a stack that can be
reverted to at the appropriate time.

---

It's important not to accidentally create empty accounts, so don't set
empty storage root in a missing account.

---

Start reporting a storage root as updated, even if the storage root is
identical.

Why? In pre-Byzantium, the storage root is reported at each transaction.
If one transaction changes the storage and then another changes it back,
then we need to treat the storage in the second transaction as changed,
even though it's identical to the start root, which is always kept as
the root at the beginning of the block.