Switch _builder -> _decoder in ChainDB, plus tests
Add tests for getting a transaction receipt, since that code path was
not exercised in testing, and caused a fatal bug when using Trinity.

Note that ChainDB uses the keyword receipt_decoder instead of
receipt_builder everywhere (and the same for transaction_decoder). At
the DB level, the logic only cares to decode from bytestrings (as
opposed to the p2p layer, which always needs to be rlp-encodable, even
if it's an opaque bytestring). Since the database only uses the "decode
from bytestring" API, in can use the "tighter" ReceiptDecoderAPI instead
of the broader ReceiptBuilderAPI. Enforcing that tighter usage seems to
make sense.