Remove json fixture cache
It was causing test failures after a VMTest move, and didn't seem to be
providing a measurable benefit (perhaps even a slowdown).

The test failure showed up as trying to load the files in the old VMTest
location in ethereum/tests. Those file paths were being loaded out of a
stale cache.