reset return data when VMError raised
VMError has a `erases_return_data` property but this never guaranteed anything. The BaseCall class somewhat accounted for this in its `__call__` method but this is sometimes not called e.g. Create(Opcode). We account for the `burns_gas` flag for the VMError on the `__exit__` call of a computation and so these changes make sure to account for the erasing of the return data for errors with the `erases_return_data` flag within this `__exit__` method as well.