Arrow Glacier hard fork / EIP-4345 support
Also updated test fixture to v10.2 to pull in the new `/DifficultyTests` and update the test suite. This had to be done by commit hash as updating the submodule was giving some issues. The commit hash can be found in the actual release notes for the new `ethereum/tests` here: https://github.com/ethereum/tests/releases/tag/v10.2