Take out eth-extra leveldb support via plyvel
It seems like binaries for some platforms are hard to come by for plyvel. It may no longer be maintained. We either subscribe to plyvel-ci or something similar or rip leveldb support out. I think at some point this may have served a more specific purpose anyhow for the Trinity client and perhaps there is little use for what seems like a decent amount of maintenance.