fix: msg.data mutability
this commit modifies `msg.data` so that it is an immutable `bytes` copy
of the input. the issue is that right now it is a view of memory, which
means the caller can trample the msg.data buffer after the call returns
(and `msg.data` gets mutated). this is not necessarily a problem for the
VM correctness, but it is an issue for integrators that inspect
`msg.data` after a call, because `msg.data` will not retain its original
value.

as a refactor, since `Computation.memory_read()` is completely dead
(and more importantly, probably a footgun), this commit also removes
`memory_read()` from the `ComputationAPI` interface as well as the
`Computation` implementation.